pyproject.toml
[project]
name = "aws-ai-practitioner"
version = "0.1.0"
description = ""
authors = [
    {name = "Aidar Darmesh",email = "darmesh.aidar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "boto3 (>=1.38.24,<2.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
